Rename ak_*.py modules -> akstr_*.py.
