+++ src/awkward/operations/str/akstr_capitalize.py
