+++ src/awkward/operations/str/akstr_center.py
