+++ src/awkward/operations/str/akstr_count_substring.py
