+++ src/awkward/operations/str/akstr_ends_with.py
