+++ src/awkward/operations/str/akstr_extract_regex.py
