+++ src/awkward/operations/str/akstr_find_substring.py
