+++ src/awkward/operations/str/akstr_find_substring_regex.py
