+++ src/awkward/operations/str/akstr_index_in.py
