+++ src/awkward/operations/str/akstr_is_alnum.py
