+++ src/awkward/operations/str/akstr_is_ascii.py
