+++ src/awkward/operations/str/akstr_is_decimal.py
