+++ src/awkward/operations/str/akstr_is_digit.py
