+++ src/awkward/operations/str/akstr_is_in.py
