+++ src/awkward/operations/str/akstr_is_lower.py
