+++ src/awkward/operations/str/akstr_is_numeric.py
