+++ src/awkward/operations/str/akstr_is_printable.py
