+++ src/awkward/operations/str/akstr_is_space.py
