+++ src/awkward/operations/str/akstr_is_upper.py
