+++ src/awkward/operations/str/akstr_join_element_wise.py
