+++ src/awkward/operations/str/akstr_length.py
