+++ src/awkward/operations/str/akstr_lower.py
