+++ src/awkward/operations/str/akstr_lpad.py
