+++ src/awkward/operations/str/akstr_ltrim.py
