+++ src/awkward/operations/str/akstr_ltrim_whitespace.py
