+++ src/awkward/operations/str/akstr_match_like.py
