+++ src/awkward/operations/str/akstr_match_substring.py
