+++ src/awkward/operations/str/akstr_match_substring_regex.py
