+++ src/awkward/operations/str/akstr_repeat.py
