+++ src/awkward/operations/str/akstr_replace_slice.py
