+++ src/awkward/operations/str/akstr_replace_substring_regex.py
