+++ src/awkward/operations/str/akstr_reverse.py
