+++ src/awkward/operations/str/akstr_rpad.py
