+++ src/awkward/operations/str/akstr_rtrim.py
