+++ src/awkward/operations/str/akstr_rtrim_whitespace.py
