+++ src/awkward/operations/str/akstr_slice.py
