+++ src/awkward/operations/str/akstr_split_pattern.py
