+++ src/awkward/operations/str/akstr_split_whitespace.py
