+++ src/awkward/operations/str/akstr_starts_with.py
