+++ src/awkward/operations/str/akstr_swapcase.py
