+++ src/awkward/operations/str/akstr_trim.py
