+++ src/awkward/operations/str/akstr_trim_whitespace.py
