+++ src/awkward/operations/str/akstr_upper.py
